refactor(github): move to empty-config. skeleton
